requirements.txt
# DO NOT include azure-functions-worker in this file
# The Python Worker is managed by Azure Functions platform
# Manually managing azure-functions-worker may cause unexpected issues

azure-functions
loguru==0.7.2
pydantic==2.6.1
pydantic-settings==2.1.0
